# CharSeq.repeat >10× speedup