original:
```java
Target            Operation   Impl                  Params  Count            Score  ±     Error    Unit  slang_persistent
CharSeqBenchmark  Repeat      slang_persistent          10      5     4,651,788.68  ±    23.81%   ops/s
CharSeqBenchmark  Repeat      slang_persistent         100      5       639,050.67  ±    20.89%   ops/s
CharSeqBenchmark  Repeat      slang_persistent        1000      5       115,828.61  ±     6.77%   ops/s
```
with doubling:
```java
Target            Operation   Impl                  Params  Count            Score  ±     Error    Unit  slang_persistent
CharSeqBenchmark  Repeat      slang_persistent          10      5    10,155,280.50  ±    12.97%   ops/s
CharSeqBenchmark  Repeat      slang_persistent         100      5     6,494,596.25  ±     4.02%   ops/s
CharSeqBenchmark  Repeat      slang_persistent        1000      5     1,204,059.76  ±     6.58%   ops/s
```
with `char[]` doubling:
```java
Target            Operation   Impl                  Params  Count            Score  ±     Error    Unit  slang_persistent
CharSeqBenchmark  Repeat      slang_persistent          10      5    13,406,385.45  ±     4.30%   ops/s                 
CharSeqBenchmark  Repeat      slang_persistent         100      5     9,117,266.10  ±     3.28%   ops/s                 
CharSeqBenchmark  Repeat      slang_persistent        1000      5     1,606,002.67  ±     4.93%   ops/s   
```