@danieldietrich 2017-02-08
Speedup looks \*great\*, thank you!